+++ bundle/language_en.properties
@@ -119,10 +119,6 @@ angal.admission.patientfolder.btn                                               
 angal.admission.patientfolder.btn.key                                                                  = S
 angal.admission.patientfolder.dicom.btn                                                                = Imaging
 angal.admission.patientfolder.dicom.btn.key                                                            = I
-angal.admission.patientfolder.visitsreport.btn                                                         = Rendez-vous
-angal.admission.patientfolder.visitsreport.btn.key                                                     = V
-angal.visits.report.dialog.title                                                                       = Rendez-vous Report
-angal.visits.report.dateinterval                                                                       = Select Date Range
 angal.admission.patientfolder.dischart.btn                                                             = Discharge Chart
 angal.admission.patientfolder.dischart.btn.key                                                         = S
 angal.admission.patientfolder.drugs.title                                                              = Drugs
