+++ bundle/language_fr.properties
@@ -115,10 +115,6 @@ angal.admission.patientfolder.btn                                               
 angal.admission.patientfolder.btn.key                                                                  = Q
 angal.admission.patientfolder.dicom.btn                                                                = Imageries
 angal.admission.patientfolder.dicom.btn.key                                                            = I
-angal.admission.patientfolder.visitsreport.btn                                                         = Rendez-vous
-angal.admission.patientfolder.visitsreport.btn.key                                                     = V
-angal.visits.report.dialog.title                                                                       = Rapport des rendez-vous
-angal.visits.report.dateinterval                                                                       = Choisir la période
 angal.admission.patientfolder.dischart.btn                                                             = Graphique de Décharge
 angal.admission.patientfolder.dischart.btn.key                                                         = E
 angal.admission.patientfolder.drugs.title                                                              = Médicaments
